+++ exporter/otlp/test/opentelemetry/exporter/otlp/exporter_test.rb
@@ -8,7 +8,6 @@ require 'test_helper'
 describe OpenTelemetry::Exporter::OTLP::Exporter do
   SUCCESS = OpenTelemetry::SDK::Trace::Export::SUCCESS
   FAILURE = OpenTelemetry::SDK::Trace::Export::FAILURE
-  TIMEOUT = OpenTelemetry::SDK::Trace::Export::TIMEOUT
 
   describe '#initialize' do
     it 'initializes with defaults' do
@@ -17,10 +16,10 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       _(exp.instance_variable_get(:@headers)).must_be_empty
       _(exp.instance_variable_get(:@timeout)).must_equal 10.0
       _(exp.instance_variable_get(:@path)).must_equal '/v1/traces'
-      _(exp.instance_variable_get(:@compression)).must_be_nil
+      _(exp.instance_variable_get(:@compression)).must_equal 'gzip'
       http = exp.instance_variable_get(:@http)
       _(http.ca_file).must_be_nil
-      _(http.use_ssl?).must_equal true
+      _(http.use_ssl?).must_equal false
       _(http.address).must_equal 'localhost'
       _(http.verify_mode).must_equal OpenSSL::SSL::VERIFY_PEER
       _(http.port).must_equal 4318
@@ -41,14 +40,29 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       assert_raises ArgumentError do
         OpenTelemetry::Exporter::OTLP::Exporter.new(compression: 'flate')
       end
-      exp = OpenTelemetry::Exporter::OTLP::Exporter.new(compression: 'gzip')
-      _(exp).wont_be_nil
       exp = OpenTelemetry::Exporter::OTLP::Exporter.new(compression: nil)
-      _(exp).wont_be_nil
+      _(exp.instance_variable_get(:@compression)).must_be_nil
+
+      %w[gzip none].each do |compression|
+        exp = OpenTelemetry::Exporter::OTLP::Exporter.new(compression: compression)
+        _(exp.instance_variable_get(:@compression)).must_equal(compression)
+      end
+
+      [
+        { envar: 'OTEL_EXPORTER_OTLP_COMPRESSION', value: 'gzip' },
+        { envar: 'OTEL_EXPORTER_OTLP_COMPRESSION', value: 'none' },
+        { envar: 'OTEL_EXPORTER_OTLP_TRACES_COMPRESSION', value: 'gzip' },
+        { envar: 'OTEL_EXPORTER_OTLP_TRACES_COMPRESSION', value: 'none' }
+      ].each do |example|
+        with_env(example[:envar] => example[:value]) do
+          exp = OpenTelemetry::Exporter::OTLP::Exporter.new
+          _(exp.instance_variable_get(:@compression)).must_equal(example[:value])
+        end
+      end
     end
 
     it 'sets parameters from the environment' do
-      exp = with_env('OTEL_EXPORTER_OTLP_ENDPOINT' => 'http://localhost:1234',
+      exp = with_env('OTEL_EXPORTER_OTLP_ENDPOINT' => 'https://localhost:1234',
                      'OTEL_EXPORTER_OTLP_CERTIFICATE' => '/foo/bar',
                      'OTEL_EXPORTER_OTLP_HEADERS' => 'a=b,c=d',
                      'OTEL_EXPORTER_OTLP_COMPRESSION' => 'gzip',
@@ -62,7 +76,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       _(exp.instance_variable_get(:@compression)).must_equal 'gzip'
       http = exp.instance_variable_get(:@http)
       _(http.ca_file).must_equal '/foo/bar'
-      _(http.use_ssl?).must_equal false
+      _(http.use_ssl?).must_equal true
       _(http.address).must_equal 'localhost'
       _(http.verify_mode).must_equal OpenSSL::SSL::VERIFY_NONE
       _(http.port).must_equal 1234
@@ -272,17 +286,17 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'retries on timeout' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_timeout.then.to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_timeout.then.to_return(status: 200)
       span_data = create_span_data
       result = exporter.export([span_data])
       _(result).must_equal(SUCCESS)
     end
 
     it 'returns TIMEOUT on timeout' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       span_data = create_span_data
       result = exporter.export([span_data], timeout: 0)
-      _(result).must_equal(TIMEOUT)
+      _(result).must_equal(FAILURE)
     end
 
     it 'returns FAILURE on unexpected exceptions' do
@@ -290,7 +304,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       logger = OpenTelemetry.logger
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_raise('something unexpected')
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_raise('something unexpected')
       span_data = create_span_data
       result = exporter.export([span_data], timeout: 1)
 
@@ -308,7 +322,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       logger = OpenTelemetry.logger
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       span_data = create_span_data
 
       Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.stub(:encode, ->(_) { raise 'a little hell' }) do
@@ -323,7 +337,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'returns TIMEOUT on timeout after retrying' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_timeout.then.to_raise('this should not be reached')
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_timeout.then.to_raise('this should not be reached')
       span_data = create_span_data
 
       @retry_count = 0
@@ -334,7 +348,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       end
 
       exporter.stub(:backoff?, backoff_stubbed_call) do
-        _(exporter.export([span_data], timeout: 0.1)).must_equal(TIMEOUT)
+        _(exporter.export([span_data], timeout: 0.1)).must_equal(FAILURE)
       end
     ensure
       @retry_count = 0
@@ -347,6 +361,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'returns FAILURE when encryption to receiver endpoint fails' do
+      exporter = OpenTelemetry::Exporter::OTLP::Exporter.new(endpoint: 'https://localhost:4318/v1/traces')
       stub_request(:post, 'https://localhost:4318/v1/traces').to_raise(OpenSSL::SSL::SSLError.new('enigma wedged'))
       span_data = create_span_data
       exporter.stub(:backoff?, ->(**_) { false }) do
@@ -355,7 +370,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'exports a span_data' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       span_data = create_span_data
       result = exporter.export([span_data])
       _(result).must_equal(SUCCESS)
@@ -366,7 +381,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       logger = OpenTelemetry.logger
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       span_data = create_span_data(total_recorded_attributes: 1, attributes: { 'a' => "\xC2".dup.force_encoding(::Encoding::ASCII_8BIT) })
 
       result = exporter.export([span_data])
@@ -381,7 +396,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'handles Zlib gzip compression errors' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_raise(Zlib::DataError.new('data error'))
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_raise(Zlib::DataError.new('data error'))
       span_data = create_span_data
       exporter.stub(:backoff?, ->(**_) { false }) do
         _(exporter.export([span_data])).must_equal(FAILURE)
@@ -389,7 +404,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'exports a span from a tracer' do
-      stub_post = stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_post = stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       processor = OpenTelemetry::SDK::Trace::Export::BatchSpanProcessor.new(exporter, max_queue_size: 1, max_export_batch_size: 1)
       OpenTelemetry.tracer_provider.add_span_processor(processor)
       OpenTelemetry.tracer_provider.tracer.start_root_span('foo').finish
@@ -399,7 +414,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'compresses with gzip if enabled' do
       exporter = OpenTelemetry::Exporter::OTLP::Exporter.new(compression: 'gzip')
-      stub_post = stub_request(:post, 'https://localhost:4318/v1/traces').to_return do |request|
+      stub_post = stub_request(:post, 'http://localhost:4318/v1/traces').to_return do |request|
         Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.decode(Zlib.gunzip(request.body))
         { status: 200 }
       end
@@ -413,8 +428,9 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'batches per resource' do
       etsr = nil
-      stub_post = stub_request(:post, 'https://localhost:4318/v1/traces').to_return do |request|
-        etsr = Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.decode(request.body)
+      stub_post = stub_request(:post, 'http://localhost:4318/v1/traces').to_return do |request|
+        proto = Zlib.gunzip(request.body)
+        etsr = Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.decode(proto)
         { status: 200 }
       end
 
@@ -429,7 +445,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     end
 
     it 'translates all the things' do
-      stub_request(:post, 'https://localhost:4318/v1/traces').to_return(status: 200)
+      stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
       processor = OpenTelemetry::SDK::Trace::Export::BatchSpanProcessor.new(exporter)
       tracer = OpenTelemetry.tracer_provider.tracer('tracer', 'v0.0.1')
       other_tracer = OpenTelemetry.tracer_provider.tracer('other_tracer')
@@ -592,8 +608,8 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
         )
       )
 
-      assert_requested(:post, 'https://localhost:4318/v1/traces') do |req|
-        req.body == encoded_etsr
+      assert_requested(:post, 'http://localhost:4318/v1/traces') do |req|
+        req.body == Zlib.gzip(encoded_etsr)
       end
     end
   end