+++ instrumentation/ruby_kafka/lib/opentelemetry/instrumentation/ruby_kafka/patches/client.rb
@@ -4,6 +4,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+require_relative '../utils'
+
 module OpenTelemetry
   module Instrumentation
     module RubyKafka
@@ -17,7 +19,9 @@ module OpenTelemetry
               'messaging.destination_kind' => 'topic'
             }
 
-            attributes['messaging.kafka.message_key'] = key if key
+            message_key = Utils.extract_message_key(key)
+            attributes['messaging.kafka.message_key'] = message_key if message_key
+
             attributes['messaging.kafka.partition'] = partition if partition
 
             tracer.in_span("#{topic} send", attributes: attributes, kind: :producer) do
@@ -35,7 +39,8 @@ module OpenTelemetry
                 'messaging.kafka.partition' => message.partition
               }
 
-              attributes['messaging.kafka.message_key'] = message.key if message.key
+              message_key = Utils.extract_message_key(message.key)
+              attributes['messaging.kafka.message_key'] = message_key if message_key
 
               parent_context = OpenTelemetry.propagation.extract(message.headers)
               OpenTelemetry::Context.with_current(parent_context) do