+++ metrics_api/lib/opentelemetry/metrics/instrument/observable_counter.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+module OpenTelemetry
+  module Metrics
+    module Instrument
+      # No-op implementation of ObservableCounter.
+      class ObservableCounter
+        # TODO
+      end
+    end
+  end
+end