+++ sdk/lib/opentelemetry/sdk/trace/tracer_provider.rb
@@ -126,8 +126,15 @@ module OpenTelemetry
         end
 
         # @api private
-        def internal_create_span(name, kind, trace_id, parent_span_id, attributes, links, start_timestamp, parent_context, instrumentation_library) # rubocop:disable Metrics/MethodLength
+        def internal_create_span(name, kind, trace_id, parent_span_id, attributes, links, start_timestamp, parent_context, parent_span, instrumentation_library) # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
+          parent_span_context = parent_span.context
+          if parent_span_context.valid?
+            parent_span_id = parent_span_context.span_id
+            trace_id = parent_span_context.trace_id
+          end
+          name ||= 'empty'
           trace_id ||= @id_generator.generate_trace_id
+          kind ||= :internal
           result = @sampler.should_sample?(trace_id: trace_id, parent_context: parent_context, links: links, name: name, kind: kind, attributes: attributes)
           span_id = @id_generator.generate_span_id
           if result.recording? && !@stopped
@@ -137,6 +144,7 @@ module OpenTelemetry
             Span.new(
               context,
               parent_context,
+              parent_span,
               name,
               kind,
               parent_span_id,